refactor module file name for consistency
