+++ src/module.indexer/indexer.spec/_testing.module.ts
@@ -2,10 +2,10 @@ import { Test, TestingModule } from '@nestjs/testing'
 import { RegTestContainer } from '@defichain/testcontainers'
 import { ConfigModule } from '@nestjs/config'
 import { ScheduleModule } from '@nestjs/schedule'
-import { DatabaseModule } from '@src/module.database/module'
+import { DatabaseModule } from '@src/module.database/_module'
 import { ModelModule } from '@src/module.model/_module'
-import { DeFiDModule } from '@src/module.defid'
-import { IndexerModule } from '@src/module.indexer/module'
+import { DeFiDModule } from '@src/module.defid/_module'
+import { IndexerModule } from '@src/module.indexer/_module'
 import { Database } from '@src/module.database/database'
 import { Indexer } from '@src/module.indexer/indexer'
 import { MemoryDatabase } from '@src/module.database/provider.memory/memory.database'
